+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/ManagementClientGenerator.cs
@@ -17,7 +17,7 @@ namespace Azure.Generator.Management
     public class ManagementClientGenerator : AzureClientGenerator
     {
         private static ManagementClientGenerator? _instance;
-        internal static ManagementClientGenerator Instance => _instance ?? throw new InvalidOperationException("ManagementClientGenerator is not loaded.");
+        internal static new ManagementClientGenerator Instance => _instance ?? throw new InvalidOperationException("ManagementClientGenerator is not loaded.");
 
         /// <summary>
         /// The Azure management client generator to generate the Azure management client SDK.