+++ eng/packages/http-client-csharp/generator/Azure.Generator/src/AzureClientGenerator.cs
@@ -8,6 +8,7 @@ using Microsoft.TypeSpec.Generator.ClientModel;
 using System;
 using System.ComponentModel.Composition;
 using System.IO;
+using Visitors;
 
 namespace Azure.Generator;
 
@@ -19,7 +20,7 @@ namespace Azure.Generator;
 public class AzureClientGenerator : ScmCodeModelGenerator
 {
     private static AzureClientGenerator? _instance;
-    internal static AzureClientGenerator Instance => _instance ?? throw new InvalidOperationException("AzureClientGenerator is not loaded.");
+    internal static new AzureClientGenerator Instance => _instance ?? throw new InvalidOperationException("AzureClientGenerator is not loaded.");
 
     /// <inheritdoc/>
     public override AzureTypeFactory TypeFactory { get; }